revng2: split off pypeline commands
Move the revng-specific pypeline command (the `quick` group, `project
init` and `pipeline run-{pipe,analysis}-native`) out of `revng2.py`
into the dedicated `pypeline_commands.py`. This leaves `revng2.py` with
registering wrappers and pathing the `pype` command-line.
